Merge commit '5f5e1db141e529a89973e6dfab95da5cb1f6a4d5' into dev

+++ tests/ffi/external_typed_data_test.dart
@@ -199,8 +199,8 @@ void testUint64Store() {
   free(ptr);
 }
 
-double maxFloat = (2 - pow(2, -23)) * pow(2, 127);
-double maxDouble = (2 - pow(2, -52)) * pow(2, pow(2, 10) - 1);
+double maxFloat = (2 - pow(2, -23)) * pow(2, 127) as double;
+double maxDouble = (2 - pow(2, -52)) * pow(2, pow(2, 10) - 1) as double;
 
 void testFloatLoad() {
   // Load
